+++ backend-rust/src/address.rs
@@ -1,6 +1,5 @@
+use crate::{graphql_api::ApiResult, scalar_types::UnsignedLong};
 use async_graphql::{ComplexObject, SimpleObject, Union};
-
-use crate::{graphql_api::ApiResult, types::UnsignedLong};
 
 #[derive(SimpleObject, Clone, serde::Serialize, serde::Deserialize)]
 pub struct AccountAddress {
