+++ backend-rust/src/lib.rs
@@ -4,6 +4,6 @@ pub mod indexer;
 pub mod address;
 pub mod decoded_text;
 pub mod router;
+pub mod scalar_types;
 pub mod transaction_event;
 pub mod transaction_type;
-pub mod types;
