+++ backend-rust/src/transaction_event/credentials.rs
@@ -1,29 +1,5 @@
-use crate::address::AccountAddress;
-use async_graphql::{InputValueError, InputValueResult, Scalar, ScalarType, SimpleObject, Value};
-
-#[derive(serde::Serialize, serde::Deserialize, derive_more::From)]
-#[repr(transparent)]
-#[serde(transparent)]
-pub struct Byte(pub u8);
-#[Scalar]
-impl ScalarType for Byte {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        let Value::Number(number) = &value else {
-            return Err(InputValueError::expected_type(value));
-        };
-        let Some(v) = number.as_u64() else {
-            return Err(InputValueError::expected_type(value));
-        };
-
-        if let Ok(v) = u8::try_from(v) {
-            Ok(Self(v))
-        } else {
-            Err(InputValueError::expected_type(value))
-        }
-    }
-
-    fn to_value(&self) -> Value { Value::Number(self.0.into()) }
-}
+use crate::{address::AccountAddress, scalar_types::Byte};
+use async_graphql::SimpleObject;
 
 #[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
 pub struct AccountCreated {
